Converted VTexPage to use bytes instead of ints to store data. Developed a hash algorithm for VTexPages that crushes its x, y and texture id components down to 4 bit numbers and then computes a hash component from all 4 values.
git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@5696 956a8c8d-885e-cb46-a0b9-a2f355b4c364

+++ Standalone/VirtualTexture/FeedbackBuffer.cs
@@ -65,7 +65,7 @@ namespace Medical
         {
             PerformanceMonitor.start("FeedbackBuffer Analyze");
             float u, v;
-            int m, t;
+            byte m, t;
             virtualTextureManager.beginPageUpdate();
             for (int slId = 0; slId < fullBitmap.Height; ++slId)
             {
